Practica03-ComputoParalelo-CUDA-en-la-Nube/src/requirements.txt
pillow
Django
djangorestframework
drf-yasg
numpy
pandas
scikit-learn
gunicorn
pillow
polls
markdown
django-filter
pyrebase4